Improved isPlayerInLobby flow
