+++ app/src/main/java/edu/agh/susgame/front/providers/web/socket/GameSocketListener.kt
@@ -1,29 +1,60 @@
 package edu.agh.susgame.front.providers.web.socket
 
+import kotlinx.coroutines.CoroutineScope
+import kotlinx.coroutines.Dispatchers
+import kotlinx.coroutines.flow.MutableSharedFlow
+import kotlinx.coroutines.flow.SharedFlow
+import kotlinx.coroutines.flow.asSharedFlow
+import kotlinx.coroutines.launch
 import okhttp3.Response
 import okhttp3.WebSocket
 import okhttp3.WebSocketListener
 import okio.ByteString
 
-class GameWebSocketListener(private val webSocketManager: WebSocketManager) : WebSocketListener() {
+class GameWebSocketListener : WebSocketListener() {
+    private val _socketOpenedFlow = MutableSharedFlow<WebSocket>()
+    private val _socketClosedFlow = MutableSharedFlow<Unit>()
+
+    private val _messagesFlow = MutableSharedFlow<String>()
+    private val _bytesFlow = MutableSharedFlow<ByteString>()
+
+    val socketOpenedFlow: SharedFlow<WebSocket> = _socketOpenedFlow.asSharedFlow()
+    val socketClosedFlow: SharedFlow<Unit> = _socketClosedFlow.asSharedFlow()
+
+    val messagesFlow: SharedFlow<String> = _messagesFlow.asSharedFlow()
+    val bytesFlow: SharedFlow<ByteString> = _bytesFlow.asSharedFlow()
+
     override fun onOpen(webSocket: WebSocket, response: Response) {
         println("WebSocket opened: ${response.message}")
+
+        CoroutineScope(Dispatchers.Main).launch {
+            _socketOpenedFlow.emit(webSocket)
+        }
     }
 
     override fun onMessage(webSocket: WebSocket, text: String) {
         println("WebSocket Receiving: $text")
-        webSocketManager.onMessageReceived(text)
+
+        CoroutineScope(Dispatchers.Main).launch {
+            _messagesFlow.emit(text)
+        }
     }
 
     override fun onMessage(webSocket: WebSocket, bytes: ByteString) {
         println("WebSocket Receiving bytes: ${bytes.hex()}")
 
-        webSocketManager.onMessageReceived(bytes.hex())
+        CoroutineScope(Dispatchers.Main).launch {
+            _bytesFlow.emit(bytes)
+        }
     }
 
     override fun onClosing(webSocket: WebSocket, code: Int, reason: String) {
         webSocket.close(1000, null)
         println("WebSocket Closing: $code / $reason")
+
+        CoroutineScope(Dispatchers.Main).launch {
+            _socketClosedFlow.emit(Unit)
+        }
     }
 
     override fun onFailure(webSocket: WebSocket, t: Throwable, response: Response?) {
